+++ tests/testthat/test-map_data_with_correspondence.R
@@ -1,3 +1,54 @@
+test_that("passing dataframe and retaining column names works", {
+  sa2_2011 <- suppressMessages(get_polygon(area = "sa2", year = 2011))
+  withr::with_seed(
+    42,
+    {
+      sa2_2011_sample <- sa2_2011[1:100,]
+
+      sa2_2011_sample_with_groups <-
+        lapply(
+          LETTERS[1:5],
+          \(x)tibble::add_column(.data = sa2_2011_sample, letter_group = x)
+        ) |>
+        (\(x) do.call("rbind", x))()
+
+      sa2_2011_sample_with_groups$test_outcome <- rnorm(nrow(sa2_2011_sample_with_groups))
+
+
+      mapped_df_with_data <- map_data_with_correspondence(
+        sa2_2011_sample_with_groups,
+        codes = sa2_code_2011,
+        values = test_outcome,
+        groups = letter_group,
+        from_area = "sa2",
+        from_year = 2011,
+        to_area = "sa3",
+        to_year = 2011,
+        value_type = "aggs"
+      )
+
+
+
+      mapped_df_without_data <- map_data_with_correspondence(
+        codes = sa2_2011_sample_with_groups$sa2_code_2011,
+        values = sa2_2011_sample_with_groups$test_outcome,
+        groups = sa2_2011_sample_with_groups$letter_group,
+        from_area = "sa2",
+        from_year = 2011,
+        to_area = "sa3",
+        to_year = 2011,
+        value_type = "aggs"
+      )
+    }
+  )
+  expect_equal(names(mapped_df_with_data), c("sa3_code_2011", "test_outcome", "letter_group"))
+  expect_equal(
+    unname(as.matrix(mapped_df_with_data)),
+    unname(as.matrix(mapped_df_without_data))
+  )
+})
+
+
 test_that("grouping works", {
   sa2_2011 <- suppressMessages(get_polygon(area = "sa2", year = 2011))
 
@@ -59,11 +110,22 @@ test_that("aggregating up SA's works", {
 test_that("mapping across SAs and editions together works", {
   sa2_2011 <- suppressMessages(get_polygon(area = "sa2", year = 2011))
 
-
-  n_sample <- 200
   withr::with_seed(
     42,
     {
+      n_sample <- 200
+      sa2_2011_sample <- dplyr::sample_n(sa2_2011, n_sample)
+      sa2_to_sa3_2011_to_2016_mapped_unit_ref_col <- map_data_with_correspondence(
+        .data = sa2_2011_sample,
+        codes = sa2_code_2011,
+        values = rnorm(n = n_sample),
+        from_area = "sa2",
+        from_year = 2011,
+        to_area = "sa3",
+        to_year = 2016
+      )
+
+
       sa2_to_sa3_2011_to_2016_mapped_unit <- map_data_with_correspondence(
         codes = sample(sa2_2011$sa2_code_2011, size = n_sample),
         values = rnorm(n = n_sample),
@@ -72,6 +134,7 @@ test_that("mapping across SAs and editions together works", {
         to_area = "sa3",
         to_year = 2016
       )
+
 
       sa2_to_sa3_2011_to_2016_mapped_aggs <- map_data_with_correspondence(
         codes = sample(sa2_2011$sa2_code_2011, size = n_sample),
@@ -84,13 +147,17 @@ test_that("mapping across SAs and editions together works", {
       )
     }
   )
-
+  expect_snapshot(sa2_to_sa3_2011_to_2016_mapped_unit_ref_col)
   expect_snapshot(sa2_to_sa3_2011_to_2016_mapped_unit)
   expect_snapshot(sa2_to_sa3_2011_to_2016_mapped_aggs)
   expect_lt(
     nrow(sa2_to_sa3_2011_to_2016_mapped_aggs),
     nrow(sa2_to_sa3_2011_to_2016_mapped_unit)
   )
+
+  expect_true(all(
+    sa2_to_sa3_2011_to_2016_mapped_unit_ref_col$sa3_code_2016 %in% asgs_2016$sa3_code_2016
+  ))
 
   expect_true(all(
     sa2_to_sa3_2011_to_2016_mapped_unit$sa3_code_2016 %in% asgs_2016$sa3_code_2016